Moves Cell css into separate file
