+++ src/Cell/Cell.css
@@ -1,0 +1,44 @@
+.crossword__cell {
+  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+  fill: white;
+  cursor: pointer;
+}
+
+.crossword__cell--highlighted {
+  fill: rgb(252, 191, 25);
+}
+
+.crossword__cell--selected {
+  fill: rgb(255, 241, 110);
+}
+
+.crossword__cell--filled {
+  fill: black;
+}
+
+/* Color for cells that are both filled and selected */
+.crossword__cell--filled.crossword__cell--selected {
+  fill: rgb(35, 69, 95);
+}
+
+.crossword__cell--number {
+  font-family: Arial, Helvetica, sans-serif;
+  fill: black;
+  font-size: 10px;
+  font-weight: 500;
+  text-anchor: start;
+}
+
+@media (max-width: 400px) {
+  .crossword__cell--number {
+    font-size: 11px;
+    transform: translateY(1px);
+  }
+}
+
+.crossword__cell--letter {
+  fill: black;
+  font-size: 17px;
+  font-weight: 500;
+  text-anchor: middle;
+}
