+++ src/PuzzleEditor.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Redirect } from "react-router-dom";
 import Context from "./Context";
-import Cell from "./Cell";
+import Cell from "./Cell/Cell";
 import Clues from "./Clues/Clues";
 import generatePattern from "./Patterns";
 import Fills from "./Fills/Fills";
